Refactor import paths in API routes and components to use a consistent structure. Updated imports in `route.ts` files and `AddToSpaceDropdown`, `CreateSpaceDialog`, and `useSpace` to streamline module resolution and enhance maintainability.
