+++ app/api/share/[hash]/route.ts
@@ -1,4 +1,4 @@
-import { prisma } from "@/app/utils/prisma";
+import { prisma } from "@/lib/prisma";
 import { NextRequest, NextResponse } from "next/server";
 
 export const GET = async (
