+++ app/api/space/route.ts
@@ -1,5 +1,5 @@
 import { authOptions } from "@/app/config/auth.config";
-import { prisma } from "@/app/utils/prisma";
+import { prisma } from "@/lib/prisma";
 import { AddToSpaceSchema } from "@/app/validators/space.validator";
 import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
